Initial attempt at linux/386 support
It passes a lot of very basic touch testing.
breakpoints can be set.
goroutines command returns sane output.
stack backtraces work.
locals seems to work.

Tests don't work, or rather they work up to the point that they hang in
proc testing.

There's a lot here that needs to be changed:

Treating TLS as a register only works when it's available from a
register. That isn't the case on 386, and the changes to make it look
like it is are really ugly. It should probably no longer be treated as a
register and just grab from a register on amd64.

I've made the 32-bit/64-bit pointer changes based off runtime.GOARCH all
over. These changes need to be consolidated/cleaned up. If cross-arch
support will ever be a thing, it shouldn't be based off runtime.GOARCH
at all but off the information in e.g. the core dump.